feat: spilt test to test chat and test flag
